marketing_agent/requirements.txt
openai
crewai
langchain
streamlit
python-dotenv
